Maintain separation between base service and service implementation.
