+++ app-modules/integration-open-ai/src/Services/OpenAiGpt4oService.php
@@ -37,18 +37,11 @@
 namespace AdvisingApp\IntegrationOpenAi\Services;
 
 use OpenAI;
-use AdvisingApp\Ai\Models\AiThread;
-use Illuminate\Support\Facades\Http;
 use AdvisingApp\Ai\Models\AiAssistant;
-use AdvisingApp\Ai\Models\AiMessageFile;
 use AdvisingApp\Ai\Settings\AiIntegrationsSettings;
-use AdvisingApp\Ai\Services\Contracts\SupportsFileUploads;
 use AdvisingApp\IntegrationOpenAi\Services\Concerns\UploadsFiles;
-use AdvisingApp\Ai\DataTransferObjects\Files\FilesDataTransferObject;
-use AdvisingApp\Ai\DataTransferObjects\VectorStores\VectorStoresDataTransferObject;
-use AdvisingApp\Ai\DataTransferObjects\VectorStoreFiles\VectorStoreFilesDataTransferObject;
 
-class OpenAiGpt4oService extends BaseOpenAiService implements SupportsFileUploads
+class OpenAiGpt4oService extends BaseOpenAiService
 {
     use UploadsFiles;
 
@@ -80,78 +73,6 @@ class OpenAiGpt4oService extends BaseOpenAiService implements SupportsFileUpload
         ]);
     }
 
-    public function retrieveFile(AiMessageFile $file): FilesDataTransferObject
-    {
-        $response = $this->client->files()->retrieve($file->file_id);
-
-        return FilesDataTransferObject::from([
-            'id' => $response->id,
-            'name' => $response->filename,
-            'status' => $response->status,
-        ]);
-    }
-
-    public function createVectorStore(array $parameters): VectorStoresDataTransferObject
-    {
-        $response = $this->client->vectorStores()->create($parameters);
-
-        return VectorStoresDataTransferObject::from([
-            'id' => $response->id,
-            'name' => $response->name,
-            'fileCounts' => get_object_vars($response->fileCounts),
-            'status' => $response->status,
-            'expiresAt' => $response->expiresAt,
-        ]);
-    }
-
-    public function retrieveVectorStore(string $vectorStoreId): VectorStoresDataTransferObject
-    {
-        $response = $this->client->vectorStores()->retrieve($vectorStoreId);
-
-        return VectorStoresDataTransferObject::from([
-            'id' => $response->id,
-            'name' => $response->name,
-            'fileCounts' => get_object_vars($response->fileCounts),
-            'status' => $response->status,
-            'expiresAt' => $response->expiresAt,
-        ]);
-    }
-
-    public function modifyVectorStore(string $vectorStoreId, array $parameters): void
-    {
-        $this->client->vectorStores()->modify($vectorStoreId, $parameters);
-    }
-
-    /**
-     * The `openai-php/client` does not current work with the `GET /vector_stores/{vectorStoreId}/files` endpoint
-     * for Azure Open AI. This is due to the expectation of a `chunking_strategy` key in the response, which Azure
-     * does not provide. An issue has been opened, but this request needs to happen without the client for now.
-     */
-    public function retrieveVectorStoreFiles(AiThread $thread, string $vectorStoreId, array $params): VectorStoreFilesDataTransferObject
-    {
-        $service = $thread->assistant->model->getService();
-
-        $response = Http::withHeaders([
-            'api-key' => $service->getApiKey(),
-            'OpenAI-Beta' => 'assistants=v2',
-            'Content-Type' => 'application/json',
-        ])
-            ->withQueryParameters(
-                $params,
-            )
-            ->get($service->getDeployment() . '/vector_stores/' . $vectorStoreId . '/files', [
-                'api-version' => $service->getApiVersion(),
-            ]);
-
-        return VectorStoreFilesDataTransferObject::from([
-            'object' => $response->json()['object'],
-            'data' => $response->json()['data'],
-            'firstId' => $response->json()['first_id'],
-            'lastId' => $response->json()['last_id'],
-            'hasMore' => $response->json()['has_more'],
-        ]);
-    }
-
     public function getApiKey(): string
     {
         return $this->settings->open_ai_gpt_4o_api_key ?? config('integration-open-ai.gpt_4o_api_key');
@@ -167,7 +88,7 @@ class OpenAiGpt4oService extends BaseOpenAiService implements SupportsFileUpload
         return $this->settings->open_ai_gpt_4o_model ?? config('integration-open-ai.gpt_4o_model');
     }
 
-    public function getDeployment(): string
+    public function getDeployment(): ?string
     {
         return $this->settings->open_ai_gpt_4o_base_uri ?? config('integration-open-ai.gpt_4o_base_uri');
     }
